src/backend/requirements.txt
# FastAPI Backend - Course Companion
# Zero-Backend-LLM Architecture (NO LLM packages allowed)

# Web Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0

# Database
sqlalchemy==2.0.25
asyncpg==0.29.0
alembic==1.13.1

# S3/R2 Storage
boto3==1.34.14
aioboto3==12.3.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Caching
redis==5.0.1

# Utilities
python-multipart==0.0.6
httpx==0.26.0
python-dateutil==2.8.2

# Development
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
black==24.1.0
ruff==0.1.13

# ============================================
# FORBIDDEN PACKAGES (Disqualification Risk)
# ============================================
# DO NOT ADD:
# - openai
# - anthropic
# - langchain
# - llama-index
# - transformers
# - sentence-transformers
# ============================================
